requirements.txt
matplotlib==3.1.2
requests==2.22.0
six==1.12.0
tensorflow_gpu==2.0.0
aiohttp==3.6.2
pyecharts==1.6.0
demjson==2.2.4
Django==3.0.2
fire==0.2.1
Appium_Python_Client==0.51
beautifulsoup4==4.9.0
facemorpher==5.2.1
Pillow==7.1.1
python_dateutil==2.8.1
tensorflow==2.1.0
